+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -65,6 +65,7 @@ public let masterRuleList = RuleList(rules: [
     NestingRule.self,
     NimbleOperatorRule.self,
     NoExtensionAccessModifierRule.self,
+    NoGroupingExtensionRule.self,
     NotificationCenterDetachmentRule.self,
     NumberSeparatorRule.self,
     ObjectLiteralRule.self,